+++ lib/instabug_flutter.dart
@@ -17,6 +17,8 @@ export 'src/modules/session_replay.dart';
 export 'src/modules/surveys.dart';
 // Utils
 export 'src/utils/instabug_navigator_observer.dart';
+export 'src/utils/private_views/instabug_private_view.dart';
+export 'src/utils/private_views/instabug_sliver_private_view.dart';
 export 'src/utils/screen_loading/instabug_capture_screen_loading.dart';
 export 'src/utils/screen_loading/route_matcher.dart';
 export 'src/utils/screen_name_masker.dart' show ScreenNameMaskingCallback;