+++ lib/src/modules/instabug.dart
@@ -16,9 +16,12 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:instabug_flutter/instabug_flutter.dart';
 import 'package:instabug_flutter/src/generated/instabug.api.g.dart';
+
+import 'package:instabug_flutter/src/generated/instabug_private_view.api.g.dart';
 import 'package:instabug_flutter/src/utils/enum_converter.dart';
 import 'package:instabug_flutter/src/utils/ibg_build_info.dart';
 import 'package:instabug_flutter/src/utils/instabug_logger.dart';
+import 'package:instabug_flutter/src/utils/private_views/private_views_manager.dart';
 import 'package:instabug_flutter/src/utils/screen_name_masker.dart';
 import 'package:meta/meta.dart';
 
@@ -152,6 +155,7 @@ class Instabug {
     BugReporting.$setup();
     Replies.$setup();
     Surveys.$setup();
+    InstabugPrivateViewApi.setup(PrivateViewsManager.I);
   }
 
   /// @nodoc