+++ pigeons/instabug_private_view.api.dart
@@ -1,0 +1,6 @@
+import 'package:pigeon/pigeon.dart';
+
+@FlutterApi()
+abstract class InstabugPrivateViewApi {
+  List<double> getPrivateViews();
+}